Homework 02

+++  Homework_Dianova_AA/homework_05/number_06.py
@@ -21,7 +21,7 @@ for line in data_file.readlines():
     zaniatija = 0
 
     for x in re.findall('\d+', line):
-        zaniatija += int(x)
+        zaniatija += float(x)
     predmety.update({predmet: zaniatija})
 
 print(predmety)
